main

+++ SomarNumero/src/somarnumero/SomarNumero.java
@@ -11,19 +11,18 @@ public class SomarNumero {
         /* int = inteiros
            double = reais
             */
-        double n1, n2, n3, r;
+        int n1, n2, n3, r;
         
         // Imprimi "Digite o primeiro número: "
         System.out.print("Digite o primeiro número: ");
         // Cria um input e atribui em n1
-        n1 = ler.nextDouble();
+        n1 = ler.nextInt();
         
         System.out.print("Digite o segundo número: ");
-        n2 = ler.nextDouble();
+        n2 = ler.nextInt();
         
         System.out.print("Digite o terceiro número: ");
-        n3 = ler.nextDouble();
-        
+        n3 = ler.nextInt();
         r = n1 + n2 + n3;
         System.out.println("A soma é igual a: " +r);
